+++ ivy/functional/backends/numpy/extensions/statistical.py
@@ -1,0 +1,18 @@
+from typing import Optional, Union, Tuple
+import numpy as np
+
+
+def median(
+    input: np.ndarray,
+    /,
+    *,
+    axis: Optional[Union[Tuple[int], int]] = None,
+    keepdims: Optional[bool] = False,
+    out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    return np.median(
+        input,
+        axis=axis,
+        keepdims=keepdims,
+        out=out,
+    )
